Finished connecting analysis functions to frontend
The word analysis functions now all work and gives a formatted text output in the main field. The only thing missing from the GUI with any significance is the search bar and connecting the table of content (enabling the user to add more books to the database from said ToC).