Search ToC and add book to DB added to GUI
A few minor changes and additions, but primarily, the search function has been added. It searches through the table of content downloaded from gutenberg and allows the user to select books found and add them to the database, which will trigger the webscraper, download the book and encode it before uploading it to the database. Setting up and downloading the table of content in the first place has yet to be added to the GUI properly (and is, iirc, the only important function missing).