Optimized code, readme for GUI & refactoring
The primary optimization would be how the book objects are initialized, but also a few small changes like implementing multi-threading in the decoder. A small analysis function was added (one where it is determined which book has the most different word usage from the rest of the selected books). The readme file for the GUI was added, taking the previous one's place (which was moved into the webscraper folder, as it mostly describes the debugging command line script found there). Some code no longer in use was removed and minor refactoring.